correcting compiling under mac, correcting README

+++ src/qt/gamepage.cpp
@@ -696,7 +696,7 @@ void GamePage::getBet()
 
                     UniValue txIn(UniValue::VOBJ);
                     txIn.pushKV("txid", txidIn);
-                    txIn.pushKV("vout", voutIdx);
+                    txIn.pushKV("vout", static_cast<int>(voutIdx));
 
                     UniValue sendTo(UniValue::VOBJ);
                     sendTo.pushKV("address", address);
